Terrain/RasterRenderer: disable slope shading depending on zoom
When the map is zoomed in or out very far, slope shading doesn't make
much sense, because the terrain resolution is not good enough or the
slope would be too tiny to be useful.  This speeds up XCSoar
conditionally, and as a side effect, it fixes #712 which was caused by
an integer overflow (dd2*dd2 could overflow).