dlgTaskEdit.cpp: merge dlgTaskManager into dlgTaskEdit
by removing dlgTaskManager, we remove an extra step.  We
clarify the UI by displaying the active task (and task point) as soon
as the user clicks the "Edit Task" option.  We have a "More" button that
shows the Declare, Browse and Save As options
Copy from dlgTaskManager.cpp the routines for
Save As, Declare, List, and Commit into dlgTaskEdit.cpp