WayPointFileZander: Added parseFlagsFromDescription() function
If the description field of a waypoint in a Zander waypoint file starts with a 1, it can be assumed that the waypoint is an airport. The SR940 is actually demanding that format, therefore we should support it too. The normal flags we supported before are just an unofficial but common addition to the standard.

This implements ticket #779