Data/Dialogs: quote the newline character in attributes
According to the XML specification, newline characters in attributes
must be translated to normal spaces:

 http://www.w3.org/TR/2000/WD-xml-c14n-20000119.html#charescaping

Our own XML parser does not do this.  This was noticed when we saw
duplicate translation strings in the PO files which did not match the
strings from the XML source (TRAC ticket #775).  The cause was that
tools/xml2po.pl uses expat, which behaves correctly.