Terrain/RasterRenderer: round quantisation_effective down
It turns out that by using ceil(), we get quantisation_effective==2
for q=1.2, and that reveals spurious artefacts by the slope shading
algorithm.  This is because the interpolation in RasterBuffer creates
some sort of moiree effect.  By rounding down quantisation_effective,
these disappear mostly, making a better fit for slope shading.

This was reported to TRAC as ticket #760.