TabBar.cpp/hpp: Create tab bar control for dialogs
This control display tabs across the top of the window and
allows the user to use mouse or keys to navigate.
It supports callbacks routines when the tabs change.
Can display tabs and / or command buttons in the tab bar.
Needs to have lazy loading support added to speed up
initial rendering.