GlueMapWindowDisplayMode: Add "None" as shift bias.
This is redundant and possibly confusing for users trying to adjust
the "Glider Position" offset and not seeing any effect (unlike for all
other "Map orientations") and certainly leads to the impression that
this is not possible for "North Up" (like before the recent changes).
But every developer has his personal favourite xcsoar behaviour quirks
he does miss - so here the option is. (BTW: This is not my favourite
quirk)