RoutePlanner: stabilise paths with minor rounding
Not sure how effective this will be, but it should make paths within 8 seconds
of each other have the same value as far as search is concerned, which
therefore will favour whichever item appears first in the search ordering,
resulting in tendency to more stable paths.

Without this patch, on occasion, almost-equal paths end up fighting for top spot.