InputEventsActions: Base lower limit of manual cruise zoom level on glider polar.
Some tasks can take advantage of a different lower zoom level limit
(e.g. paragliding). Instead of fixing the lower limit to 5km base
it on the V min of the configured polar and limit cruise zoom to
the distance that can be minimally flown in 2 minutes of straight flight.

Implements trac enhancement request #542.