WayPointRenderer: Minor rendering tweaks for runways.
Width of runway (wx, wy calculation):
Since the width was calculated with an rounded int offset from the center
this could result in nearly 2 pixel rendering differences, i.e.
Old: iround(2.6) + iround(2.6) => 3 + 3 = 6 runway width.
     iround(2.4) + iround(2.4) => 2 + 2 = 4 runway width.
New: iround(2.6 + 2.6) => 5, iround(2.4 + 2.4) => 5;
This can result in a slightly off center rendering which for width is barley
visible. However this calculation has the sideeffect of a larger error for the
length components. Reduce these by forcing these to be a multiple of 2.