pnlTaskManagerClose.cpp: implement OnTabReClick()
If the user clicks on the Close tab while not flying and no changes are
made to the task, it displays the Close screen (it does not automatically
close).  That behavior was needed to provide the option to go to the
TaskCalculator.
The ReClick function simply allows clicking Close twice will close the Task Manager.
(Note: it still closes in one click while flying because while flying, the TMgr was
opened via the Calculator).