AbortTask: use hook system to perform terrain intersection tests.
This partially resolves #736.

This mechanism will make it easy to filter out other destinations
(e.g. airspace etc).  Uses RoutePolars system for fast terrain intersections.

The implementation could be further optimised by rolling into one lock the
intersection tests, but for now this is functional.

Intersections are only performed for final glide checks, since if we have to climb,
then it is assumed that obstacles can be avoided.

Note that the AlternateTask sorting does not pay any attention to terrain intersections yet.