OLC: Add SIS-AT contest optimisation.
Implements rules as per
http://sis-at.streckenflug.at/2011/pdf/sis-at-ausschreibung.pdf

Note that this is similar to OLC Classic, and optimised as such.
What is affected is purely the scoring points; according to the
documentation the path optimisation is to maximise the 6 leg
free distance (therefore equivalent to OLC Classic).

This closes ticket #448.
Note: this has not been well tested.