OS/Clock: Remove __attribute__(const) from MonotonicClockMS()
__attribute(const)__ is used to mark functions which have no side effect
and which are not effected by global state. With no parameters functions
always must return the same value and can therefore be subject to aggressive
CSE. Clearly this will not work for MonotonicClockMS() since it is supposed
to return global state (the tick counter). For much the same reason
__attribute__(pure) will also not work reliable since the compiler can't
notice the change of global scope data since no C global variables of the
same execution thread are involved.