TaskFile, TaskStore, NEWS.txt: display names of CUP tasks
Need to parse the task names of a CUP file when we count the
number of tasks before displaying the list.  The string is already read
into a variable, so not performance degradation is noticeable.
NEWS.txt: add SeeYou task files note
Add dummy TaskFile::~TaskFile() {} to RunAnalysis.cpp
to resolve linker reference