default.xci, altair.xci: show only one Task menu button
The one button pulls up the task manager, the first page is
the task calculator.  Only available in the menu
when not in Abort mode.
Make it visible if not in Abort/Goto mode.
Does not matter if waypoint file is valid because
user can load a task w/out a waypoint file.