Menu: implement next/previous arm buttons
Code is implemented to enable use of nextarm and previousarm buttons and events.
The way these buttons work is as follows.  For tps not requiring arming,
the next/previous buttons are labeled and function as before.

For tps requiring arm, the "next" button becomes "Arm start" or "Arm turn",
and function to arm the advance (if not already armed).  If already armed,
the button becomes "next".  Thus, two button presses will skip the tp.

For tps requiring arm, the "previous" button functions as "previous" if the
current tp is not armed, or if armed, will disarm the turnpoint and read "Disarm start"
or "Disarm turn" as appropriate.

This is not yet hooked up to the *.xci input specification.