NMEA/Info: Provide pressure altitude (baro altitude with QNH=1013.25 hpa as reference)
This distinguishes between baro altitude and pressure altitude. To calculate
a correct baro altitude from a pressure altitude, a valid QNH setting is needed (and vice versa).

So we set the altitude which is given from the source and do not calculate
the other one if QNH is not valid.