AirspaceRoute: exclude airspaces inside origin and destination
Previously, planning will fail if the origin or destination are inside
airspace.  This patch excludes those airspaces from the local store used by the
planning system so valid paths will be produced even in those cases.

It is up to the pilot whether he/she wants to fly to a target inside airspace,
or is already in airspace!