StartPoint: fix failure of find_best_start in worst cases
Depending on relative position of aircraft and the perimeter, the
solver would fail due to inability of the ZeroFinder to handle wraparound
of the parametric boundary functions [0,1).

This patch solves that issue fixing #975.  Note that some switching between
optimal points on either side (behind or in front) of perimeter if
behind the area may occur still due to the actual numerical ambiguity
about which is the superior path.  That's just life.