Poco/RWLock: remove WIN32 implementation, use Mutex instead
In the 6.1 branch, reproducible deadlocks can be observed with the
WIN32 RWLock implementation.  It is safe to say that RWLocks are not
worth the trouble, because none of the "production" targets use it -
both Windows CE and Android fall back to a plain Mutex.  This patch
removes the WIN32 implementation.