TaskStats: correct instantaneous task speed for target changes
The task remaining effective instantaneous speed is a good statistic
only if the targets do not change.  If they do, then spikes result
in the speed values, as reported in #581.

The targets change due to AAT optimisation, and so the effect
of these spikes can be reduced by subtracting out the changes in
task distance.