GlideComputerAirData: Provide correct values for Average30s and NettoAverage30s
The old implementation essentially initialised the values to the
Brutto/Netto vario value on detection of a reset condition and then did
not update the values during the next 30 seconds.

The new calculation will use the samples obtained until now for
average calculation even if the full 30 seconds have not yet been sampled.
Also removed the LowPassFilter since the average calculation should be
sufficient.