RasterTile: track progress via step counter
Exit when number of steps equals the max steps, not by checking exact
value.  This actually fixes a possible bug where the terrain is low-res
(making it impossible to land on the target), and also allows for subsampling
speedup in a later patch.

It should also speed things up slightly because we are only comparing one value, not two.