DeviceBlackboard: SetStartupLocation() is no-op if GPS connected
Don't set the location when there is a real GPS connected or replay is
running.  This eliminates the "Flying" check in dlgWayPointDetails,
Components.cpp and UtilsSettings.cpp.

A slight semantic change is introduced: when a real GPS has provided a
location or replay mode is active, then the new home location is
ignored, even if no takeoff was registered yet.