NMEA/Info: move pressure formulas to BasicComputer::Fill()
Move calculations to a "computer" class.  This is more preparation for
the NMEA_INFO split, and allows us to remove the *Origin enums now.

At the same time, we adopt a new policy: when we know the current
pressure and the QNH, then the external "true" altitude setting is
discarded and a new one is calculated, because the external device
which provided it may have a different QNH setting.