dlgConfiguration: Remove alternate rendering style of final glide bar
The only difference between the default and the alternate rendering style
is the label position of the "required altitude gain" for completing
the task. Since alternate renders the label right of the glide bar it
intrudes further into the map area and can cause bigger conflicts with
other display elements.

Implements trac enhancement request #1099.