InputEvents: trigger gps on UI events if no valid fix
If no valid fix, the engine and several display elements do not update
(e.g. setting mc does nothing to infobox display, changing display modes does
nothing etc).

This patch generates a synthetic gps update event if the gps is not
connected or is connected but doesnt yet have a fix.

This closes #918.