GlueMapWindowDisplayMode: smooth incorporation of thermal center
Previously, the screen center was either at the aircraft or at the thermal center,
if too far, clicking from one to the other which is distracting.

This patch feeds in the thermal center as an interpolated value so
the aircraft is always visible but screen center biases towards the thermal center.
This prevents switching between aircraft and thermal center.