Form/List: correct key handling behaviour
This patch makes two changes necessary for allowing navigation of
dialogs incorporating Lists for non-pointer devices e.g. Altair.

Firstly, XCSoar uses the standard for all properties that left-right adjusts values,
up-down selects previous/next control.  So this patch swaps usage of left-right and up-down
which were page up, page down and previous item, next item respectively.

Secondly, handling of page up, page down at the extremity of the list needs to not trap the key
so the OS can catch this as focus previous, focus next respectively.

The result of this patch is that navigation using keys only works on all forms,
especially the task manager.  Previously it was impossible to select an item in the task list and
then select the "move up" field, for example.

This should not have any noticeable effect when using the mouse or touchscreen.

This closes #499.