TaskFileSeeYou.cpp: fix error w/ SeeYou's default OZ.
When See You exports a task in a .cup file, OZs in the task that are the same as SeeYou's
default OZ are exported with no OZ information.  When XCSoar reads these files, those OZs should be
created using the XCSoar defaults.
Currently, they are being created with 500/500 cylinder with inner radius.
Now they are created based on XCSoar's Task Defaults in configuration.