MapWindowSymbols: dont draw best cruise track arrow in climb mode
If in purely climb mode (not WAITCRUISE), dont draw the best cruise track arrow.
It's not required when fully circling.  As soon as the glider turns away (WAITCRUISE or CRUISE)
becomes active, the arrow will be visible.

The arrow is only required for accurate direction during cruise, so presenting
it in climb mode only distracts the user and clutters the screen.