MapDrawHelper: Fixed m_canvas.alpha_blend() call
We can assume that m_buffer in some cases might be larger than m_canvas due
to the MapWindow::on_resize() function. If that happens m_buffer is drawn with the
right projection but when using AlphaBlend() it is stretched to fit the size of m_canvas.
(see http://msdn.microsoft.com/en-us/library/dd183351(v=vs.85).aspx)

If we only copy/AlphaBlend() with the rectangle of m_canvas nothing will be stretched
and the original projection remains.

This should fix ticket #1095