SerialPort: Listen for EV_BREAK too in WaitCommEvent() function
In the StopRxThread() function Flush() is used (which calls
the PurgeComm() function) before closing the port.
On some Windows PCs PurgeComm() never returns because it is
waiting for a lock that the WaitCommEvent() holds.

Since PurgeComm() is used with the BREAK parameters, we can
make WaitCommEvent() listen for the EV_BREAK event too and
continue the loop which checks is_stopped() as soon as it is detected.

This fixes ticket #1137