Graphics: color adjustment for non-alert traffic
Use a desaturated blue (same family as mark and thermal source)
as this is neutral (does not indicate goodness as green does), and
does not stand out as much as previous green did.

This therefore makes the orange and red alert colours pop more.

Note that FlarmTrafficWindow already draws in priority layers so
non-alert filled targets will not obscure alert targets.