make: Added LINK macro to fix CCache usage on Windows
When ccache was called for linking a binary it returned without waiting for the
spawned process to return. When the strip step was called right afterwards it failed
because the binary wasn't linked yet.
The linking step will now be done without using ccache on windows to prevent
the step from returning too early.