PolarGlue: Display Polar message box only if polar was invalid.
The PolarGlue will show a Message Box showing the polar is wrong
even if you don't have a polar set. This means the first thing a new
user sees is an error "Polar has invalid coefficients".
This now checks the profile, and if it is empty, does not show the message
but still resets the polar. If the polar is wrong, it still shows the message.