InfoBoxes/Content/Task: Display frequency in infobox comment area if it exists.
Frequency is an important information when a landing on a foreign
airfield is imminent due to height. Having to search for this information
in a paper map or trying to look it up in the WP detail dialog during
a high workload situation is potentially dangerous. Showing it in the
next WP infobox makes the frequency easily available for use.