OpenGL/RawBitmap: use GL_UNSIGNED_SHORT_5_6_5 for the texture
This solves the terrain rendering problem on the Samsung Galaxy S II
GT-I9100 (with Mali-400 GPU), ticket #1228.

The "type" parameter of glTexImage2D() describes the type of the
buffer that was passed; when the buffer is NULL, the argument should
be ignored.  It seems however that the Mali-400 driver remembers the
value, and returns GL_INVALID_OPERATION on the next glTexSubImage2D()
call.

Thanks to Triadis for donating a Samsung Galaxy S II for debugging.