FinishPoint: Reverted fix for finish cylinders
See ticket #1385

When using the usual finish lines an optimization step is applied.
Since the finish line is perpendicular to the previous task leg, the optimized
target point should be exactly the reference point of the finish line.
Instead of this a different point is chosen before even starting the task.

In one instance I noticed that the optimized point was even at the end of the
finish line.