Device/Driver: don't pass DerivedInfo to PutQNH()
Only the Vega driver needs it, but only because the Vega expects other
setting to be sent along with the QNH.  This patch postpones the QNH
setting to the next OnSysTicker() call.  As a side effect, we can
remove the lock from DeviceBlackboard::SetQNH(), which fixes TRAC
ticket #1389 (assertion failure due to double lock).