SearchPointVector: remove bugged "convex" nearest point code
That code path was completely wrong, because it assumes that the
nearest point must be on one sides of the nearest vertex.  However,
with a very stretched polygon, the nearest point can be at the back
side, while the nearest side (that includes the nearest point) is not
considered.

For known convex polygons, there is no obvious simplification or
speedup we can use; the one that is deleted by this patch was a
fallacy.