Components.cpp: call check_duplicate_waypoints() when loading default task
After creating an OrderedTask, it is required that the taskpoints are
deduped and added to the waypoints structure.  This was not happening
in task_load_default() and task_load(), both of which called task_create() followed
immediately by commit().
Because check_duplicate_waypoints requires ScopeSuspendAllThreads,
we did not want to add that code to ProtectedTaskManager so that PTM is more reusable.
So, we removed task_load() and task_load_default() and required the clients to call task_create() followed
by the check_duplicate_waypoints() followed by task_commit().

Removed task_load() and task_load_default() and modified clients who call them to
call appropriate equivalent code with check_duplicate_waypoints().