TaskLeg: always travel from current location to current destination
In leg_vector_remaining(), always calculate the remaining vector from
the current aircraft position, no matter if a "previous" turnpoint
exists.  The previous turnpoint isn't used anyway, so the origin check
didn't make a lot of sense.  This calculates the vector from the
aircraft to the start point properly, and of course the time needed
for that.

It fixes ridiculous speeds that are caused by a huge distance vector
and a small time (>2000m/s for a start point that was a few hundred
kilometers away).  These wrong speeds caused miscalculations in
airspace arrival times (TRAC ticket #1392).

As a side effect, we can remove the special case for current start
point from OrderedTask::glide_solution_remaining().

Another side effect: the final glide bar now shows the real height
required for finishing the task, including the height required to
reach the start point.  Previously, the correct height requirement was
only shown after the start point was skipped.