MapTaskManager.cpp: remove calls to check_duplicate_waypoints
Removed call to check_duplicate_waypoints() from remove_from_task() and
from insert_in_task() (called by dlgWaypointSelect.cpp).
When editing a task by adding or removing waypoints via the map, there is
no need to call check_duplicate waypoints because all the waypoints
displayed on the map are by definition already in the waypoint file.
Also, the task has already been deduped against the waypoint file becase we
call check_duplicate_waypoints() whenever any new waypoints are added to
a task or when the waypoint tree changes:
(1) at startup when Default.tsk is loaded
(2) when the waypoint file is changed or modified
(3) when any saved task is loaded
So, all waypoints shown on the map and all waypoints in the task have already
been deduped.  And any new waypoint being added to the task from the map
is by definition already an exact match of one in the waypoint structure.