Profile: save profile changes to the file that was loaded
This patch ultimately removes the ability to make profiles
incremental: instead of loading xcsoar-registry.prf first and then the
file chosen by the user on top, we now only load the user's personal
file.

When saving the profile, it is saved back to the user's file.

The only remaining reference to xcsoar-registry.prf is that it's the
default file name for new XCSoar installations.