ExternalLogger.cpp: overwrite IGC file if exists
If the IGC file is downloaded and it already
exists with this filename, it should overwrite it
and keep the correct name and flight number.
Prior behavior was to create a second IGC
file and increment the flight number in the file
name if a file with that name existed.  The
flight number represents the order the flights
were flown, not the order they were downloaded.
Tested on CAI302.
This (and the previous commit to
ExternalLogger.cpp) fix trac #1381