TaskDefaultsConfigPanel: make radius/length label dynamic
When the Start Type is changed to a "Line"
then it should have a "Gate Width" property.  This
reportedly translates better to other languages.
And when the Start Type is set to any sector type
or cylinder type, then it should have a "Radius"
property.  Same for Finish Type.
Dynamically change the label display based
on the type of start / finish.  Also update help text.
Affects:
1) Taskpoint Edit screen in the task manager
2) Task Defaults Turnpoint Types screen in the Configuration settings
3) Task point list in the task manager
Fixes #1272