FastMath: round to signed integer in NATIVE_TO_INT()
Casting to an unsigned integer gives some odd rendering artefacts on
Android (with FPU support) when negative angles are passed to
Canvas::segment(), as described in TRAC ticket #1297.  We need to be
able to deal with negative angles, and casting to unsigned was never
necessary - the bit-wise AND with 0xfff corrects everything after the
integer cast.