TabBar.cpp: fix landscape bug
If landscape screen and
SetClientOverlapTabs(true)
was not explicitly set by the
client, nor was "Vertical" set
in the XML, then the tab's client
windowswere drawn in the wrong place.
The only impact was on dlgStatus
which has code to correct the problem.
Next commit will remove the fix code
from dlgStatus.