doc/html: move documentation to the developer manual
Much of doc/html/ has been redundant since the TeX manual was created,
and most of the rest is outdated.  This patch attempts to move much of
the rest to the developer, and deletes the directory "doc/html/".  All
future effort should concentrate on the TeX manual.

The new TeX contents are not formatted well, but that's better than
not having the text at all.