Screen/ButtonWindow: restore the "Enter" key in dialogs
When the button has the keyboard focus, and the user presses "Enter"
(VK_RETURN), call on_click().  This fixes TRAC ticket #1459.

The WIN32 dialog manager does not do this by default; it would require
us to move the "default push button" around with the keyboard focus.
That seems a little overkill for how little we want it to do.