Data: Use red graphics for TESTING builds
The only difference between TESTING and non-TESTING was the icon that
was clicked before the application was started. There was no indication where
the user is running TESTING or not after that point. We should clearly indicate
that the user is running a potentially unstable version of XCSoar by using
more of the testing build color inside of the application.