Android: acquire "Vibrate" permission
A bug report on the Android market has revealed an interesting detail:
on newer Android versions, you need the "VIBRATE" permission to add
notifications, or your application will crash from a
"SecurityException: Requires VIBRATE permission".  Since we are
planning to add vibrate support anyway, let's get the permission now,
to work around that problem.

It is unclear why the vibration was added to our "ongoing event"
notification, because we did not request it.