Makefile: set default symbol visibility to "hidden"
The Android port is implemented as a shared library which gets loaded
by the Java VM via JNI.  As a side effect, all internal symbols are
public, and will be in the symbol table.  That's quite a bad thing,
because the symbol table is extremely large.

By making all symbols "hidden" by default, and only explicitly
exporting the JNI symbols, the Android binary shrinks by 350 kB.