AtmosphericPressure: FindQNH() gets a raw pressure altitude
Don't assume the altitude that is passed to FindQNH..() is adjusted by
current QNH setting - this didn't make sense, because if the QNH
setting were known, AutoQNH would not need to calculate it.  I believe
this is a regression from the time when we introduced the difference
between "barometric altitude" (above sea level) and "pressure
altitude" (above 1013.25 hPa).