RoutePolar: fix angle calculations
The formulas for converting projected X/Y values to a polar index were
utterly broken.  This caused big differences between straight arrival
height and "ReachFan" arrival height when the wind was strong, because
the head/tail wind was not considered correctly (TRAC #1531).

The integer-only formula is now replaced with a floating point one.
The cost of that is acceptable.