ElementStat: add attribute vector_remaining
This is the true unmodified remaining vector from the aircraft to the
next waypoint (task point; target).  This replaces
solution_remaining.vector, which was unreliable: whent he head wind
was very strong, the glide solver fails, leaving a zero vector in the
GlideResult, and thus the InfoBox showed a distance of 0km.
Sometimes, it would show a partial solution, which was even more
confusing.

Thanks to Andreas Pfaller and Tobias Bieniek who spent many hours
analysing and debugging this issue.  This fix is based on their
findings.