AndroidIOIOUartPort: use IOIO Manager
Modify AndroidIOIOUartPort to use the global
ioio_manager object.

Change IOIOHelper.java to connect to the ioio board
with open() via AddClient() instead of openUart().

Delete constructor in IOIOHelper.java and move code
to open().

Change IOIOHelper to not disconnect with the ioio board
with close() in the cpp destructor because it disconnects
with close() via RemoveClient()

Remove unnecessary code in close() to close all uarts
because they should already all be closed.