TrafficWidget: new widget to replace the FLARM traffic dialog
This is a draft Widget that will replace the modal FLARM traffic
dialog.  Traffic will be shown as a "page", replacing the map, and the
user can still interact with InfoBoxes, the menu and other dialogs.

The implementation is incomplete.  There are no visible buttons to
return to the map and for zooming.  You cannot open the details of
FLARM traffic.

This code is currently disabled.  To activate it, call the event
"Traffic show".