add configure script
The script (configure) will give build instructions, and, if run on a Mac, run
the appropriate configure script:

The script in build/configure-mac.sh guesses the GCC variant to use and verifies it.
It produces a (partial) Makefile, build/local-config.mk, which is included if
available.  It sets a default (TARGET) and defines the compile/link binaries.
(This .mk file is added to .gitignore)

GCC is found if named gcc-4.6 (through 4.4) in PATH, or in
/usr/local/bin/ (gcc-4.6 or gcc), or simply gcc in PATH.
The gcc that ships with OS X is currently at version 4.2, which is insuffient
to understand -std=gnu++0x.