dlgConfiguration: fix cancel hardware button
We removed the Cancel button from the configuration
dialog button some time ago.  The developers manual
was updated in the developers manual saying that a dialog should either
have Save and Cancel buttons, or just a Close button.
If only the Close button exists, it would act like a Save button.
Therefore, the Hardware button for Cancel should act like
the Close button and save the settings on dlgConfiguration
because there is no Cancel button.
It was exiting the dialog without saving.  Now it saves.