Remote stick support
The Quick Menu that supports easy navigation with cursor keys. The
menu buttons on the Quick Menu can be customized by creating an *.xci
file that contains a list of menu buttons. The *.xci file should be
placed in the folder XCSoarData. There can be at most 32 buttons on
the Quick Menu which may be displayed on multiple pages depending on
the device screen size.