GlueMapWindow: Removed circling mode special case for trail rendering
When doing wave soaring and returning to a spot of a constant wave,
this behaviour prevented the use of the previous flight track to catch
the wave again. The user can still configure the behaviour if the full
track is too long.

Also it is now possible to hide the trail also in circling mode,
which was fixed to 5min (short trail) before.

This fixes ticket #1565.