ThermalAssistantWidget: a non-modal version of the TA
This new Widget replaces the old modal TA dialog.  A menu hack is used
to show a "close" button.  This is the first Widget that gets
integrated into the UI, because it is a very simple widget with no
configuration.  It will be used to experiment with the Widget / Page
concept; we need to explore how we will allow the user to switch back
to the map.