Widget: add Leave() virtual method, enforce
Add Leave() method to widget with API rule that Leave()
shall be called prior to calling Hide().
Leave() returns a bool that allows the caller to decide
whether to proceed with calling Hide().
Enforce the API rule (in ManagedWidget, TabbedControl
and MainWindow) where Hide() is called.
Create NULLWidget::Leave() that returns true.