Trace: remove "delta_list_iterator" attribute
This attribute was a shortcut to quickly erase elements, but it had a
major problem: it used the TraceDelta class before it was fully
defined.

The performance advantage of the "delta_list_iterator" attribute could
not be demonstrated in a benchmark.  By removing this attribute, 5%
memory is saved, with no disadvantage.