Trace: use std::multiset
One property of std::set is that it implicitly discards duplicate
values.  The Trace doesn't (shouldn't) have duplicate values, but if
it eventually does, it will silently drop one TraceDelta, leading to a
segmentation fault.  On the other hand, eliminating the old value is a
tiny bit of overhead which is avoided with this patch.