GlueMapWindow: multi-touch drag triggers pan
Users expect to have an easy way to just pan the map.  Dragging with
one finger is already reserved for gestures.  This patch adds an
experimental change: when dragging with two fingers, pan mode is
activated automatically.  This seems somewhat consistent, because we
may add two-finger zooming at the same time (not implemented yet).

Still it's not intuitive enough: how will users know that they should
use two fingers?  Is two-finger dragging accessible in a cockpit?
Let's experiment some more with this.