TaskMacCready: don't consider minimum heights twice
In the end of glide_solution(), the minimum heights and the climb
height were applied again.  The goal of this code is not clear, and it
seems to introduce severe calculation errors in the "total" altitude
difference when the finish height is above the current height.  Based
on a patch by Andreas Pfaller, and fixes TRAC ticket #1593.