MacCready: use new local task in solve_sink()
A "virtual" task was created with a huge offset, but it was never
used; that offset was subtracted from the result, though, always
rendering wrong values.

This was not a real problem, because TaskStats::glide_required was
used only in the debug programs.  It was not used by XCSoar at all.