dlgInfoBoxAccess: remove "InfoBox:" from dialog title
Display the descriptive name of the InfoBox in the
titlebar of the Dialog.  Don't show the word "InfoBox:"
which is arguably accurate but of low value
to the pilot.  Of more value is the title.
A rule of thumb is that the less reading a design
requires, the more effective the design.