ShowInfoBoxPicker: add itemhelp
Enable itemhelp display below the list of InfoBoxes
as the user scrolls through the list of info boxes.
Leave the "Help" button which is sometimes needed to
show very lengthy descriptions of a few InfoBoxes.
So, the "Help" buton will generally show the same
information as the itemhelp window below the list,
unless there is too much text to fit in itemhelp,
in which case "Help" button will show an entire page
of help text.