ListPicker: fix initial position when item help is enabled
Was not always displaying the initial_item in the visible window
when the list_control was initially displayed and itemhelp was
enabled.

If itemhelp is enabled, then the initial scroll position of the
scrollbar needs to be set (with SetCursorIndex()) after the
list_control is resized to the smaller dimensions that allow
for the display of the itemhelp below it.