dlgInfoBoxAccess: call Save() after dialog closes
If the dialog closes for any reason, we always save
the data.  Now, we call Save() so code can be
placed in the Save() methods of the widgets.
Noteably, if the modeless InfoBox dialog disappears
if a click elsewhere on the screen, this will
save the data.
ToDo: implement Save() in the InfoBox dialog content