Android/InternalGPS: fix race condition
Patch changes the update() method in InternalGPS.java (and an
analogous method in a new java class for baro and other sensors) to be
synchronous, i.e. update() does not return until the main thread has
updated the sensor/GPS subscriptions. This fixes a race condition that
occurred during XCSoar shutdown, where the InternalGPS (.cpp) object
destructor was causing InternalGPS (.java) object to request to be
unsubscribed from the GPS, then deleting the Java object before the
main thread could service the request.  Usually a segfault would occur
and the XCSoar icon/"XCSoar is running" would persist in the
notification pulldown.