GlideSolvers/MacCready: fix optimise_glide function
class MacCreadyVopt does not take into account the wind drift when
mc is not zero and this lead to a wrong speed to fly. The error
originates in function GlideResult::CalcVInvSpeed, since the distance
used in the calculation is not corrected with the drift that would result
after a climb.

When mc is not zero the glide should be calculated at the optimum
speed correspondig to the MacCready value. Only when trying to
maximize the glide relative to ground the speed to fly is different
to the mc=0 speed to fly.

- /src/Engine/GlideSolvers/MacCready.cpp: MacCreadyVopt changed
  to find the speed to maximize glide when mc=0.

- MacCready::solve_glide: added time_virtual calculation.

- /src/Engine/GlideSolvers/GlideResult.cpp:
  Member function GlideResult::CalcVInvSpeed no longer needed.

- /src/Engine/Task/Tasks/AbortTask.cpp: no need to call any function
  to calculate time_virtual.