List.cpp: move ActivateItem() to OnMouseUp()
Create DragMode::CURSOR to track when the selected item is dragged
and move call to ActivateItem() into OnMouseUp().

* In OnMouseUp, call ActivateItem instead of from OnMouseDn and call it
only if drag_mode==CURSOR and the mouse is horizontally over the list.
This way, if the mouse is dragged off the list to left or right and released
it won't activate the callback.
* In OnMouseMove change drag_mode from CURSOR to SCROLL if vertical
dragging exceeds 1/5 cursor height from the MouseDown location.
* In OnMouseMove, perform scroll if drag_mode==CURSOR.
* In OnMouseDn, move setting of drag_y so it is set when drag_mode is
set to either CURSOR or SCROLL.
* In OnMouseDn, call the kinetic.MouseDown for both drag_mode==CURSOR and
drag_mode==SCROLL.
* In OnMouseDn, track the mouse down vertical location relative to the
window in a new property called drag_y_window to be used for detecting
vertical movement beyond a threshhold amount.