KineticManager.cpp: fix sporadic failure to scroll kinetically
It was a bug to call the MouseMove from MouseUp because the MouseMove
event was fired prior to the MouseUp from the same position.  This resulted
in calling MouseMove with the same x twice in a row, resulting in a velocity "v" of
0, causing the kinetic scroll to fail.  This happened on slower systems more because
of the filter in MouseMove the prevents it from calculating v twice within 15ms.

I tested on PC, Streak 5 and Streak 7 and DROIDX.  For the PC with the super-fast CPU,
dt was always below 15 ms, so the scrolling worked perfectly.  For the Streak 5 and
DROIDX, dt was usually below 15ms, so the scrolling usually worked because the second
MouseMove was filtered out and ignored.  For the Streak 7, dt was never below 15 ms, so
the scrolling always failed.
fixes bug#1679