KineticManager: fix "bounce" during fast scroll
It was an error not to set the "v" velocity to 0
on MouseDown.  The circumstances that created an error was
during a rapid scroll, if v was fast (e.g. 4.0) and the dt was 500ms,
so it was visually slowing down, and a finger "bounces" on the screen.
The finger "bounce" triggers MouseDown then MouseUp with no
MouseMove because the finger does not move, just down and up.
However, it updates the clock, so the GetPosition thinks that a new
scroll has starting with the original velocity of 4.0, and the scroll
would visually accelerate.  The correct behavior is to immediately
freeze the scroll on MouseDown, and only scroll if the new drag
scrolls the list.