AbstractTask: remove GlidePolar reference attribute
Pass the GlidePolar when needed.  The problem here is that the
AbstractTask will refer to a GlidePolar object that must be managed by
the caller all the time.  It is not clear who updates it and how it
should be protected (multi-threading).  This removes a lot of
GlidePolar references where they were never really used.